feat: 🎸 implement API base controller and give it auth power
