+++ app/controllers/api/stays_controller.rb
@@ -1,4 +1,4 @@
-class Api::StaysController < ApplicationController
+class Api::StaysController < Api::BaseController
   def index
     @stays = Stay.all
     render json: @stays
